Refactor monolithic defaults into modules.
